+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/github/lint_github_reporter.dart
@@ -14,14 +14,15 @@ const _deprecationMessage =
 ///
 /// **Note:** this reporter is deprecated and should not be used.
 /// Consider using Dart Code Metrics GitHub Action instead.
-class LintGitHubReporter
-    extends GitHubReporter<LintFileReport, SummaryLintReportRecord<Object>> {
+class LintGitHubReporter extends GitHubReporter<LintFileReport,
+    SummaryLintReportRecord<Object>, void> {
   const LintGitHubReporter(IOSink output) : super(output);
 
   @override
   Future<void> report(
     Iterable<LintFileReport> records, {
     Iterable<SummaryLintReportRecord<Object>> summary = const [],
+    void additionalParams,
   }) async {
     if (records.isEmpty) {
       return;