+++ lib/src/analyzers/unused_files_analyzer/reporters/report_params.dart
@@ -1,0 +1,6 @@
+/// Represents additional reporter params
+class ReportParams {
+  final bool deleteUnusedFiles;
+
+  const ReportParams({this.deleteUnusedFiles = false});
+}