+++ lib/src/analyzers/unused_l10n_analyzer/unused_l10n_analyzer.dart
@@ -25,7 +25,7 @@ class UnusedL10nAnalyzer {
 
   /// Returns a reporter for the given [name]. Use the reporter
   /// to convert analysis reports to console, JSON or other supported format.
-  Reporter<UnusedL10nFileReport, void>? getReporter({
+  Reporter<UnusedL10nFileReport, void, void>? getReporter({
     required String name,
     required IOSink output,
   }) =>