+++ lib/src/cli/commands/check_unused_files_command.dart
@@ -2,6 +2,7 @@
 
 import 'dart:io';
 
+import '../../analyzers/unused_files_analyzer/reporters/report_params.dart';
 import '../../analyzers/unused_files_analyzer/unused_files_analyzer.dart';
 import '../../config_builder/config_builder.dart';
 import '../models/flag_names.dart';
@@ -30,6 +31,7 @@ class CheckUnusedFilesCommand extends BaseCommand {
     final folders = argResults.rest;
     final excludePath = argResults[FlagNames.exclude] as String;
     final reporterName = argResults[FlagNames.reporter] as String;
+    final deleteFiles = argResults[FlagNames.deleteFiles] as bool;
 
     final config = ConfigBuilder.getUnusedFilesConfigFromArgs([excludePath]);
 
@@ -40,12 +42,19 @@ class CheckUnusedFilesCommand extends BaseCommand {
       sdkPath: findSdkPath(),
     );
 
+    if (deleteFiles) {
+      _analyzer.deleteAllUnusedFiles(unusedFilesResult);
+    }
+
     await _analyzer
         .getReporter(
           name: reporterName,
           output: stdout,
         )
-        ?.report(unusedFilesResult);
+        ?.report(
+          unusedFilesResult,
+          additionalParams: ReportParams(deleteUnusedFiles: deleteFiles),
+        );
 
     if (unusedFilesResult.isNotEmpty &&
         (argResults[FlagNames.fatalOnUnused] as bool)) {
@@ -57,6 +66,7 @@ class CheckUnusedFilesCommand extends BaseCommand {
     _usesReporterOption();
     addCommonFlags();
     _usesExitOption();
+    _usesDeleteUnusedFiles();
   }
 
   void _usesReporterOption() {
@@ -83,6 +93,16 @@ class CheckUnusedFilesCommand extends BaseCommand {
         help: 'Treat find unused file as fatal.',
 // TODO(dkrutrkikh): activate on next major version
 //        defaultsTo: true,
+      );
+  }
+
+  void _usesDeleteUnusedFiles() {
+    argParser
+      ..addSeparator('')
+      ..addFlag(
+        FlagNames.deleteFiles,
+        abbr: 'd',
+        help: 'Delete all unused files.',
       );
   }
 }