+++ lib/src/cli/models/flag_names.dart
@@ -28,4 +28,5 @@ class FlagNames {
   static const l10nClassPattern = 'class-pattern';
 
   static const fatalOnUnused = 'fatal-unused';
+  static const deleteFiles = 'delete-files';
 }