+++ lib/src/reporters/models/html_reporter.dart
@@ -8,7 +8,8 @@ import 'file_report.dart';
 import 'reporter.dart';
 
 /// HTML-doc reporter
-abstract class HtmlReporter<T extends FileReport, S> extends Reporter<T, S> {
+abstract class HtmlReporter<T extends FileReport, S, P>
+    extends Reporter<T, S, P> {
   static const String id = 'html';
 
   @protected
@@ -21,6 +22,7 @@ abstract class HtmlReporter<T extends FileReport, S> extends Reporter<T, S> {
   Future<void> report(
     Iterable<T> records, {
     Iterable<S> summary = const [],
+    P? additionalParams,
   }) async {
     if (records.isEmpty) {
       return;