+++ test/src/cli/commands/check_unused_files_command_test.dart
@@ -22,6 +22,9 @@ const _usage = 'Check unused *.dart files.\n'
     '\n'
     '    --[no-]fatal-unused                           Treat find unused file as fatal.\n'
     '\n'
+    '\n'
+    '-d, --[no-]delete-files                           Delete all unused files.\n'
+    '\n'
     'Run "metrics help" to see global options.';
 
 void main() {